Dockerize the go backend which should fix things
The error in question was nginx not behaving as expected, which I think
could have been caused by it being in docker when the backend wasn't?
It is fixed now.

Another advantage is that when we inevitably make this run on a linux
server, host.docker.internal won't work in the nginx.conf, so now I can
use go-backend instead

Now, the go code is inside the go-backend directory, and the root is for
things like the docker-compose.yml